+++ ft_calloc.c
@@ -6,18 +6,18 @@
 /*   By: toespino <toespino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/22 12:30:46 by toespino          #+#    #+#             */
-/*   Updated: 2025/11/04 18:51:36 by toespino         ###   ########.fr       */
+/*   Updated: 2026/01/04 23:47:43 by toespino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_calloc(size_t nmemb, size_t size)
+void	*ft_calloc(uint64_t nmemb, uint64_t size)
 {
-	size_t	mas;
-	void	*res;
+	uint64_t	mas;
+	void		*res;
 
-	if (nmemb > 0 && size > ((size_t)-1) / nmemb)
+	if (nmemb > 0 && size > ((uint64_t)-1) / nmemb)
 	{
 		return (0);
 	}
