+++ ft_memchr.c
@@ -6,16 +6,16 @@
 /*   By: toespino <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/15 18:27:00 by toespino          #+#    #+#             */
-/*   Updated: 2025/10/22 16:57:29 by toespino         ###   ########.fr       */
+/*   Updated: 2026/01/04 23:52:52 by toespino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memchr(const void *s, int c, size_t n)
+void	*ft_memchr(const void *s, int32_t c, uint64_t n)
 {
 	char			*dest;
-	size_t			i;
+	uint64_t			i;
 	unsigned char	*temp;
 
 	i = 0;
