+++ ft_memcmp.c
@@ -6,15 +6,15 @@
 /*   By: toespino <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/19 14:40:38 by toespino          #+#    #+#             */
-/*   Updated: 2025/10/22 13:59:21 by toespino         ###   ########.fr       */
+/*   Updated: 2026/01/04 23:55:12 by toespino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_memcmp(const void *s1, const void *s2, size_t n)
+int32_t	ft_memcmp(const void *s1, const void *s2, uint64_t n)
 {
-	size_t			i;
+	uint64_t		i;
 	unsigned char	*sb1;
 	unsigned char	*sb2;
