+++ ft_memmove.c
@@ -6,13 +6,13 @@
 /*   By: toespino <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 18:35:08 by toespino          #+#    #+#             */
-/*   Updated: 2025/10/17 14:31:22 by toespino         ###   ########.fr       */
+/*   Updated: 2026/01/04 23:58:21 by toespino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memmove(void *dest, const void *src, size_t n)
+void	*ft_memmove(void *dest, const void *src, uint64_t n)
 {
 	unsigned char		*ptr1;
 	const unsigned char	*ptr2;
